Changed casting to float from int. WP

+++ elegantnumbers.py
@@ -1,6 +1,6 @@
 print("Enter your numbers to be summed up:")
-x = int(input())
-y = int(input())
+x = float(input())
+y = float(input())
 z = x + y
 output = "{} plus {} equals {}!"
